anno: fix clippy lints, doc collision, update README version
- Remove unnecessary f64 cast in relation_extraction.rs
- Use enumerate() instead of indexing in glirel/onnx.rs
- Replace iter().any(|t| *t == x) with contains(&x) in qa_regression
- Add [lib] doc = false to root crate to resolve doc name collision
- Update README dependency version to 0.3.6

+++ crates/anno/src/backends/inference/relation_extraction.rs
@@ -1241,7 +1241,7 @@ pub fn extract_relation_triples_simple(
                 confidence *= distance_penalty;
 
                 // Incorporate entity confidence.
-                confidence *= (head.confidence + tail.confidence) as f64 / 2.0;
+                confidence *= (head.confidence + tail.confidence) / 2.0;
 
                 if confidence < config.threshold as f64 {
                     continue;
